+++ components/dashboard/blogBasicInfo/BlogUrl.tsx
@@ -7,22 +7,19 @@ import BLOG_BASIC_INFO_DATA from '@/constants/blogBasicInfoData';
 const BlogUrl = () => {
   return (
     <>
-      <BlogInfoTitle>블로그 설정</BlogInfoTitle>
-      <>
-        <BlogUrlTitle>블로그 주소</BlogUrlTitle>
-        <BlogUrlText>{BLOG_BASIC_INFO_DATA.myBlog.blogUrl}</BlogUrlText>
-      </>
+      <BlogUrlTitle>블로그 주소</BlogUrlTitle>
+      <BlogUrlText>{BLOG_BASIC_INFO_DATA.myBlog.blogUrl}</BlogUrlText>
     </>
   );
 };
 
 export default BlogUrl;
 
-const BlogInfoTitle = styled.p`
-  ${({ theme }) => theme.fonts.Heading1};
-  margin-top: 6rem;
-  color: ${({ theme }) => theme.colors.grey_900};
-`;
+// const BlogInfoTitle = styled.p`
+//   ${({ theme }) => theme.fonts.Heading1};
+//   margin-top: 6rem;
+//   color: ${({ theme }) => theme.colors.grey_900};
+// `;
 
 const BlogUrlTitle = styled.p`
   ${({ theme }) => theme.fonts.Body2_Semibold};
